Move doc-cli tests to new dir
